refactor(mission): extract Mission.ReviewOps (review lifecycle + gate)
Continues the god-module decomposition (Mission 7014→6565→6578, ReviewOps
235). ReviewOps owns get_review* / get_review_with_context /
list_reviews_for_session / latest_review_for_task / review_gate_status /
decision_hygiene_findings / execution_ready? / submit_review /
respond_review + hygiene helpers. Mission defdelegates to it; normalize/
memory/prompt helpers stay in Mission and are called back via Mission.*.
Restores the maybe_record_prompt_outcome catch-all (plan-only clause was
leaked during the split).

mix precommit: 2252/0 (1 excluded), CI guard passed.
